meta_ad_analyzer/requirements.txt
apify-client==1.8.1
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
plotly==5.22.0
openpyxl==3.1.2